Add sideEffects directive to package.json
The sideEffects directive can help module bundlers remove unused imports. By asserting this package doesn't have side effects, it's safe to remove `import "extra-array"` and its contents, if none of the imports are used.

Here's an explanatory document in case you aren't familiar. It's in the webpack documentation, but bundlers in general adopt this convention.

https://webpack.js.org/guides/tree-shaking/

Thanks for this project, I'll work a lot on arrays, and I like that so many utilities are available in one place!